Configs are changed to toml
1) The config is toml and the interface for the configManger has been
changed. Previously, a config could have been retrieved with a '.'
separated string (core.db.path). In the current version, the config path
has to be passed as arguments to the getConfig
function (getConfig("core", "db", "path")).
2) Same changes also apply to the setConfig.
3) Changed the structure of the config files that network file generator
writes. Previously, nfg would write 2 files for substrate and 2 for
VNRs. Now, it only writes one file including all variables. The
structure of the written file is also identical to that of config.toml.
4) A minor fix to python_utilities: these were causing issue with
inclusion of Python.h, if the code was being compiled without setting
the WITH_FNSS_SUPPORT flag. Therefore, made their compilations dependent
on this flag.