Colorized CLI and Reading Defaults from toml
TOML file path can now be given as an argument to CLI and we wont need
the baked in path. I have not removed that yet to have a default
fallback.

Added color to the CLI and improved the options.

The default values are read and set when parsing the CLI arguments from
the given config file path.

Added a locking mechanism to the configmanager isntead of automatically
locking when a config is read.